Put factory function into companion object, require application id
